extract GanttRowBars component
