+++ frontend/src/components/gantt/NewGanttChart.vue
@@ -69,83 +69,17 @@
 					>
 						<!-- Row content with relative positioning -->
 						<div class="gantt-row-content">
-							<!-- SVG container for bars in this row -->
-							<svg
-								class="gantt-row-bars"
-								:width="totalWidth"
-								height="40"
-								xmlns="http://www.w3.org/2000/svg"
-							>
-								<g
-									v-for="bar in ganttBars[index]"
-									:key="bar.id"
-								>
-									<!-- Main bar -->
-									<rect
-										:x="getBarX(bar)"
-										:y="4"
-										:width="getBarWidth(bar)"
-										:height="32"
-										:rx="4"
-										:fill="getBarFill(bar)"
-										:stroke="getBarStroke(bar)"
-										:stroke-width="getBarStrokeWidth(bar)"
-										:stroke-dasharray="!bar.meta?.hasActualDates ? '5,5' : 'none'"
-										class="gantt-bar"
-										@pointerdown="handleBarPointerDown(bar, $event)"
-									/>
-									
-									<!-- Left resize handle -->
-									<rect
-										:x="getBarX(bar) - 3"
-										:y="4"
-										:width="6"
-										:height="32"
-										:rx="3"
-										fill="var(--white)"
-										stroke="var(--primary)"
-										stroke-width="1"
-										class="gantt-resize-handle gantt-resize-left"
-										@pointerdown="startResize(bar, 'start', $event)"
-									/>
-									
-									<!-- Right resize handle -->
-									<rect
-										:x="getBarX(bar) + getBarWidth(bar) - 3"
-										:y="4"
-										:width="6"
-										:height="32"
-										:rx="3"
-										fill="var(--white)"
-										stroke="var(--primary)"
-										stroke-width="1"
-										class="gantt-resize-handle gantt-resize-right"
-										@pointerdown="startResize(bar, 'end', $event)"
-									/>
-									
-									<!-- Task label with clipping -->
-									<defs>
-										<clipPath :id="`clip-${bar.id}`">
-											<rect
-												:x="getBarX(bar) + 2"
-												:y="4"
-												:width="getBarWidth(bar) - 4"
-												:height="32"
-												:rx="4"
-											/>
-										</clipPath>
-									</defs>
-									<text
-										:x="getBarTextX(bar)"
-										:y="24"
-										class="gantt-bar-text"
-										:fill="getBarTextColor(bar)"
-										:clip-path="`url(#clip-${bar.id})`"
-									>
-										{{ bar.meta?.label || bar.id }}
-									</text>
-								</g>
-							</svg>
+							<GanttRowBars
+								:bars="ganttBars[index]"
+								:total-width="totalWidth"
+								:date-from-date="dateFromDate"
+								:day-width-pixels="DAY_WIDTH_PIXELS"
+								:is-dragging="isDragging"
+								:is-resizing="isResizing"
+								:drag-state="dragState"
+								@barPointerDown="handleBarPointerDown"
+								@startResize="startResize"
+							/>
 						</div>
 					</VikunjaStyledGanttRow>
 				</div>
@@ -160,7 +94,6 @@ import {useRouter} from 'vue-router'
 
 import { useGlobalNow } from '@/composables/useGlobalNow'
 import {getHexColor} from '@/models/task'
-import {colorIsDark} from '@/helpers/color/colorIsDark'
 
 import type {ITask, ITaskPartialWithId} from '@/modelTypes/ITask'
 import type {DateISO} from '@/types/DateISO'
@@ -171,6 +104,7 @@ import VikunjaStyledGanttChart from '@/components/gantt/styled/VikunjaStyledGant
 import VikunjaStyledGanttRow from '@/components/gantt/styled/VikunjaStyledGanttRow.vue'
 import Loading from '@/components/misc/Loading.vue'
 import GanttVerticalGridLines from '@/components/gantt/GanttVerticalGridLines.vue'
+import GanttRowBars from '@/components/gantt/GanttRowBars.vue'
 
 import {MILLISECONDS_A_DAY} from '@/constants/date'
 import {useWeekDayFromDate} from '@/helpers/time/formatDate'
@@ -304,43 +238,6 @@ function transformTaskToGanttBar(t: ITask): GanttBarModel {
 	return bar
 }
 
-// Computed properties for dynamic bar positions during drag/resize
-const getBarX = computed(() => (bar: GanttBarModel) => {
-	if (isDragging.value && dragState.value?.barId === bar.id) {
-		const originalX = computeBarX(dragState.value.originalStart)
-		const offset = dragState.value.currentDays * DAY_WIDTH_PIXELS
-		return originalX + offset
-	}
-	if (isResizing.value && dragState.value?.barId === bar.id && dragState.value.edge === 'start') {
-		const newStart = new Date(dragState.value.originalStart)
-		newStart.setDate(newStart.getDate() + dragState.value.currentDays)
-		return computeBarX(newStart)
-	}
-	return computeBarX(bar.start)
-})
-
-const getBarWidth = computed(() => (bar: GanttBarModel) => {
-	if (isResizing.value && dragState.value?.barId === bar.id) {
-		if (dragState.value.edge === 'start') {
-			const newStart = new Date(dragState.value.originalStart)
-			newStart.setDate(newStart.getDate() + dragState.value.currentDays)
-			const originalEndX = computeBarX(dragState.value.originalEnd)
-			const newStartX = computeBarX(newStart)
-			return Math.max(0, originalEndX - newStartX)
-		} else {
-			const newEnd = new Date(dragState.value.originalEnd)
-			newEnd.setDate(newEnd.getDate() + dragState.value.currentDays)
-			const originalStartX = computeBarX(dragState.value.originalStart)
-			const newEndX = computeBarX(newEnd)
-			return Math.max(0, newEndX - originalStartX)
-		}
-	}
-	return computeBarWidth(bar)
-})
-
-const getBarTextX = computed(() => (bar: GanttBarModel) => {
-	return getBarX.value(bar) + 8
-})
 
 /**
  * Update ganttBars when tasks change
@@ -421,64 +318,6 @@ function onRowFocus() {
 	// Handle row focus if needed
 }
 
-// Direct SVG bar rendering functions
-function computeBarX(startDate: Date) {
-	const x = (startDate.getTime() - dateFromDate.value.getTime()) / (1000*60*60*24) * DAY_WIDTH_PIXELS
-	return x
-}
-
-function computeBarWidth(bar: GanttBarModel) {
-	const diff = (bar.end.getTime() - bar.start.getTime()) / (1000*60*60*24)
-	const width = diff * DAY_WIDTH_PIXELS
-	return width
-}
-
-function getBarFill(bar: GanttBarModel) {
-	// For tasks with actual dates
-	if (bar.meta?.hasActualDates) {
-		// Use task color if available
-		if (bar.meta?.color) {
-			return bar.meta.color
-		}
-		// Default to primary color if no task color
-		return 'var(--primary)'
-	}
-	
-	// For tasks without dates, use gray
-	return 'var(--grey-100)'
-}
-
-function getBarStroke(bar: GanttBarModel) {
-	if (!bar.meta?.hasActualDates) {
-		return 'var(--grey-300)' // Gray for dashed border
-	}
-	return 'none'
-}
-
-function getBarStrokeWidth(bar: GanttBarModel) {
-	if (!bar.meta?.hasActualDates) {
-		return '2'
-	}
-	return '0'
-}
-
-function getBarTextColor(bar: GanttBarModel) {
-	const black = 'var(--grey-800)'
-	
-	// For tasks without actual dates, use dark text on gray background
-	if (!bar.meta?.hasActualDates) {
-		return black
-	}
-	
-	// For tasks with actual dates
-	if (bar.meta?.color) {
-		// Use colorIsDark to determine text color based on background
-		return colorIsDark(bar.meta.color) ? black : 'white'
-	}
-	
-	// Default for primary color background (white text)
-	return 'white'
-}
 
 // Double-click and drag detection
 let lastClickTime = 0
@@ -806,27 +645,6 @@ const dateIsToday = computed(() => (date: Date) => {
 	width: 100%;
 }
 
-.gantt-row-bars {
-	position: absolute;
-	top: 0;
-	left: 0;
-	pointer-events: none;
-	z-index: 4;
-	
-	:deep(.gantt-bar) {
-		pointer-events: all;
-		cursor: grab;
-		
-		&:hover {
-			opacity: 0.8;
-		}
-	}
-
-	:deep(text) {
-		pointer-events: none;
-		user-select: none;
-	}
-}
 
 // Ensure rows have minimum height and proper styling
 :deep(.gantt-row) {
@@ -844,44 +662,4 @@ const dateIsToday = computed(() => (date: Date) => {
 	}
 }
 
-// SVG bar styling
-.gantt-bar {
-	cursor: grab;
-	
-	&:hover {
-		opacity: 0.8;
-	}
-	
-	&:active {
-		cursor: grabbing;
-	}
-}
-
-.gantt-bar-text {
-	font-size: .85rem;
-	pointer-events: none;
-	user-select: none;
-}
-
-// Resize handles
-:deep(.gantt-resize-handle) {
-	cursor: col-resize !important;
-	opacity: 0;
-	transition: opacity 0.2s ease;
-	pointer-events: all; // Ensure they receive pointer events
-	
-	&:hover {
-		opacity: 1;
-	}
-}
-
-// Show resize handles on bar hover
-:deep(g:hover) .gantt-resize-handle {
-	opacity: 0.8;
-	
-	&:hover {
-		opacity: 1;
-		cursor: inherit; // Use the specific cursor defined above
-	}
-}
 </style>
