+++ frontend/src/components/gantt/styled/VikunjaStyledGanttBar.vue
@@ -5,7 +5,7 @@
 		:timeline-end="timelineEnd"
 		:on-move="onMove"
 		:on-double-click="onDoubleClick"
-		as="rect"
+		as="svg"
 	>
 		<template #default="{ dragging, selected, focused }">
 			<rect
@@ -38,7 +38,13 @@ import GanttBarPrimitive from '../primitives/GanttBarPrimitive.vue'
 import type {GanttBarModel} from '@/composables/useGanttBar'
 import {colorIsDark} from '@/helpers/color/colorIsDark'
 
-const props = defineProps<{ model:GanttBarModel; timelineStart:Date; timelineEnd:Date; onMove:(id:string,start:Date,end:Date)=>void; onDoubleClick?:(model:GanttBarModel)=>void }>()
+const props = defineProps<{ 
+	model:GanttBarModel; 
+	timelineStart:Date; 
+	timelineEnd:Date; 
+	onMove:(id:string,start:Date,end:Date)=>void; 
+	onDoubleClick?:(model:GanttBarModel)=>void 
+}>()
 const PIXELS_PER_DAY = 30
 const MILLISECONDS_PER_DAY = 1000 * 60 * 60 * 24
